Implement a DebateXML exporter
DebateXML is an interchange and archival format for debate tournaments
of multiple styles, including BP.

This commit adds a quick DebateXML exporter function to tournament tab-
masters, which is also available after the tournament closes. The link
is available under the Tournament drop-down with the Tabbie1/DebReg
integrations.

A helper function was also added to the tournament model to fetch all
questions at once, rather than just a type thereof. Further, novice
status was not implemented in the exporter as it does not seem to be
complete.

Schema: TabbycatDebate/dta-spec#4